Trying to fix navigation in the uploaded versions of the game
Cleaned out the game files so that there is only one html file and it is named index.html.

Updated the navigateTo function in the js file again to try to work out the issues preventing proper direction when the game is uploaded online. Also added in navigation protection to keep it from being overridden. This is to prevent external directions from breaking the game when it is uploaded online.