Updated the cypress tests to search like a user then added logging
We rearranged the word finder tests to search the grid first then switch to placement data searches as a back up. After that we added logging to track how much grid searching vs placement data searching was happening so we could tell if searching like a user was failing.

Also I dug up the kethaneum puzzles file which just holds the full story data for the luminos book. I put that in with the game files so it wouldn't get lost again.