Fixed a bug with placing words that was breaking the game
There was a word placement bug that got much worse following the modularization of the architecture. So we added in pseudo-random number generator to help with this and changed how it attempts placement before switching to the deterministic way of placing. However we will need to remove the console logging once the game is more stable.