Added in new buttons, hooked up main menu button, and refactored the code
We added in the rest of the buttons on the library nav panel and hooked up the main menu button to take the player to the main menu.

However, we noticed that the navigation was in the module bootstrap instead being in the navigation.js file. So we copied over the proper functions into the correct module. But we have not yet removed the unnecessary functions from the moduleBootstrap file.